% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logit_link}
\alias{logit_link}
\title{calculate the logistic link function}
\usage{
logit_link(x)
}
\arguments{
\item{x}{numeric vector. Values to calculate the logistic link function for}
}
\description{
calculate the logistic link function for one or more observations
}
\author{
Janek Thomas
}

