% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logitreg}
\alias{logitreg}
\title{Calculate a logistic regression}
\usage{
logitreg(design, response, ...)
}
\arguments{
\item{design}{numeric matrix. Design matrix containing the observations, first
column should be constant 1 as intercept}

\item{response}{numeric vector. Response vector of 1s and 0s, should have the
same length as the number of rows in design.}

\item{...}{Further parameters passed to \code{\link[stats]{optim}}.}
}
\value{
A list with estimated coefficients, fitted propabilities and
original data
}
\description{
Calculate a logistic regression for given data. Uses
\code{\link[stats]{optim}} with the Broyden-Fletcher-Goldfarb-Shanno (BFGS)
algorithm  for parameter estimation
}
\author{
Janek Thomas
}

